LLMs/writing-prompts/requirements.txt
openai==2.4.0
python-dotenv==1.1.1